Resolve current state of NaN
Fixes #1

Caught by setting a conditional breakpoint in render function with a simplified version of the loop.
```js
! this.props.data.map(function (src, i) { return _this.state.current === i ? 'block' : 'none'; }).join('').match('block')
```

This is because `e.deltaY` can be `0` or `-0`. which means we have a divide-by-0 leading to `delta = NaN`  and a permanent state in the render function where nothing gets out of hidden.